New script that preprocesses input mrcs files into format for cryoDRGN training
Preprocessing steps include:
* windowing images in real space
* inverting data sign
* computing FFT (technically HT)
* downsampling (optional)
* symmetrizing HT (DxD -> (D+1)x(D+1) image)

Loads and saves the data in chunks to avoid OOM issues. Also uses multiprocessing
to parallelize FFT computation.

This is usually done after loading the data in cryodrgn train_vae (to avoid storing
multiple copies of the data). We now modularize this step since data preprocessing
is the memory bottleneck for training large datasets and to set up the inputs
in a standard format for chunked data loading.